Add files via upload
Dr. John Snow (1813-1858) was a famous British physician and is widely recognized as a legendary figure in the history of public health and a leading pioneer in the development of anesthesia. Some even say one of the greatest physicians of all time.  This is not to be confused with the popular character from Game of Thrones.

Dr. Snow is now also recognized as one of the founders of modern epidemiology.  Dr. Snow used a map of central London to determine the cause of Cholera, which was a water pump.  Unfortunately, other practicing physicians in England did not believe him for many years.  This is a data story about Dr. Snow's data story.